Query tile count directly
Due to the `INSERT OR IGNORE` statement, duplicated tiles would count towards the in-memory tile count, leading to tiles getting purged before they were required. The other solution to this was to either run a uniquing query before the `INSERT` or remove the `OR IGNORE` and handle the error, but both of those seemed like they'd be less performant than simply fetching the tile count up front.