Touch tile synchronously
For reasons I'm not entirely sure of, for some of our tile sources, the asynchronous touching of the tiles caused the cache to miss more often than it hit, which in turn caused tiles to flash constantly. Turning the tile touch synchronous seems to solve the issue without affecting performance adversely.